+++ src/parser/inblock/autolink.ts
@@ -1,7 +1,5 @@
 ﻿import { AutolinkParser } from '../inblock';
 import { union } from '../../combinator';
-import { hashtag } from './autolink/hashtag';
 
 export const autolink: AutolinkParser = union([
-  hashtag,
 ]);