+++ src/parser/inline.test.ts
@@ -73,6 +73,24 @@ describe('Unit: parser/inline', () => {
       assert.deepStrictEqual(inspect(parser(' @a')), [[' ', '<a class="account" rel="noopener">@a</a>'], '']);
     });
 
+    it('hashtag', () => {
+      assert.deepStrictEqual(inspect(parser('#a#')), [['#', 'a#'], '']);
+      assert.deepStrictEqual(inspect(parser('#a#b')), [['#', 'a#', 'b'], '']);
+      assert.deepStrictEqual(inspect(parser('#a')), [['<a class="hashtag" rel="noopener">#a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('#a\nb\n#c\n[#d]')), [['<a class="hashtag" rel="noopener">#a</a>', '<span class="linebreak"> </span>', 'b', '<span class="linebreak"> </span>', '<a class="hashtag" rel="noopener">#c</a>', '<span class="linebreak"> </span>', '<a href="#index:d" rel="noopener">d</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('##a')), [['##', 'a'], '']);
+      assert.deepStrictEqual(inspect(parser('a#b')), [['a#', 'b'], '']);
+      assert.deepStrictEqual(inspect(parser('a\n#b')), [['a', '<span class="linebreak"> </span>', '<a class="hashtag" rel="noopener">#b</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('a\\\n#b')), [['a', '<br>', '<a class="hashtag" rel="noopener">#b</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('*a*#b')), [['<em>a</em>', '<a class="hashtag" rel="noopener">#b</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('((a))#b')), [['<sup class="annotation">a</sup>', '<a class="hashtag" rel="noopener">#b</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[[a]]#b')), [['<sup class="authority">a</sup>', '<a class="hashtag" rel="noopener">#b</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('[#a')), [['[', '<a class="hashtag" rel="noopener">#a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser(' [#a')), [[' ', '[', '<a class="hashtag" rel="noopener">#a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('|#a')), [['|', '<a class="hashtag" rel="noopener">#a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser(' #a')), [[' ', '<a class="hashtag" rel="noopener">#a</a>'], '']);
+    });
+
   });
 
 });